requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.12
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.0
MarkupSafe==3.0.2
narwhals==1.38.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
protobuf==6.30.2
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pymongo==4.12.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.2
watchdog==6.0.0
websockets==15.0.1

